api/requirements.txt
flask==2.3.3
gunicorn==21.2.0
google-cloud-storage==2.14.0
google-cloud-pubsub
